tag がロード後に表示されないバグの修正
