+++ src/app/[service]/_components/BookmarkGrid.tsx
@@ -1,7 +1,7 @@
 import { Bookmark } from '@/types/bookmark';
 import { getVisibleBookmarks } from '@/lib/bookmark';
 import { fetchTagList } from '@/actions/tagList';
-import BookmarksSection from './BookmarksSection';
+import BookmarkGridClient from './BookmarkGridClient';
 
 interface Props {
   bookmarks: Bookmark[];
@@ -17,19 +17,18 @@ const BookmarkGrid: React.FC<Props> = async ({
   const visible = getVisibleBookmarks(bookmarks, editable);
   const entries = await Promise.all(
     visible.map(async (bm) => {
-      const { tags } = await fetchTagList(bm.id);
-      return [bm.id, tags] as const;
+      const { tags, tagListId } = await fetchTagList(bm.id);
+      return [bm.id, { tags, tagListId }] as const;
     }),
   );
-  const bookmarkTagMap: Record<string, string[]> = Object.fromEntries(entries);
-
-  const tags = Array.from(new Set(entries.flatMap(([, tags]) => tags))).sort();
+  const bookmarkTagDataMap = Object.fromEntries(entries);
+  const tags = Array.from(new Set(entries.flatMap(([, d]) => d.tags))).sort();
 
   return (
-    <BookmarksSection
+    <BookmarkGridClient
       bookmarks={visible}
+      bookmarkTagDataMap={bookmarkTagDataMap}
       tags={tags}
-      bookmarkTagMap={bookmarkTagMap}
       editable={editable}
       servicePath={servicePath}
     />
