+++ src/app/[service]/_components/BookmarkGridClient.tsx
@@ -1,96 +1,75 @@
 'use client';
 
-import { useState, useMemo } from 'react';
-import { Bookmark } from '@/types/bookmark';
+import React, { useState } from 'react';
 import { sortBookmarksByDate, groupBookmarksByYearMonth } from '@/lib/bookmark';
+import { Bookmark } from '@/types/bookmark';
 import { YearMonthCard } from './YearMonthCard';
 import { BookmarkCard } from './BookmarkCard';
 import EmptyBookmark from './EmptyBookmark';
 
-interface TagInfo {
-  tagListId: string;
-  tags: string[];
-}
-
-interface Props {
+type Props = {
   bookmarks: Bookmark[];
-  tagMap: Record<string, TagInfo>;
+  bookmarkTagDataMap: Record<string, { tags: string[]; tagListId: string }>;
+  tags: string[];
   editable: boolean;
   servicePath?: string;
-}
+};
 
 export default function BookmarkGridClient({
   bookmarks,
-  tagMap,
+  bookmarkTagDataMap,
+  tags,
   editable,
   servicePath,
 }: Props) {
-  const [selectedTags, setSelectedTags] = useState<string[]>([]);
+  const [selectedTag, setSelectedTag] = useState('ALL');
 
-  const allTags = useMemo(() => {
-    const s = new Set<string>();
-    bookmarks.forEach((b) =>
-      (tagMap[b.id]?.tags || []).forEach((t) => s.add(t)),
-    );
-    return Array.from(s).sort();
-  }, [bookmarks, tagMap]);
-
-  const filtered = useMemo(() => {
-    if (selectedTags.length === 0) return bookmarks;
-    return bookmarks.filter((b) =>
-      selectedTags.some((t) => tagMap[b.id]?.tags.includes(t)),
-    );
-  }, [bookmarks, tagMap, selectedTags]);
+  const filtered =
+    selectedTag === 'ALL'
+      ? bookmarks
+      : bookmarks.filter((b) =>
+          bookmarkTagDataMap[b.id].tags.includes(selectedTag),
+        );
 
   const sorted = sortBookmarksByDate(filtered);
   const grouped = groupBookmarksByYearMonth(sorted);
 
-  const toggleTag = (tag: string) =>
-    setSelectedTags((prev) =>
-      prev.includes(tag) ? prev.filter((t) => t !== tag) : [...prev, tag],
-    );
-
-  if (sorted.length === 0) return <EmptyBookmark />;
+  if (!sorted.length) return <EmptyBookmark />;
 
   return (
-    <div>
-      <div className="flex flex-wrap gap-2 mb-6">
-        {allTags.map((tag) => (
-          <button
-            key={tag}
-            type="button"
-            className={`px-3 py-1 rounded-full border transition ${
-              selectedTags.includes(tag)
-                ? 'bg-blue-500 text-white border-blue-500'
-                : 'bg-gray-100 text-gray-700 border-gray-300'
-            }`}
-            onClick={() => toggleTag(tag)}
-          >
-            {tag}
-          </button>
-        ))}
+    <div className="mx-auto max-w-screen-xl px-4">
+      <div className="mt-6 mb-4 flex justify-center">
+        <select
+          value={selectedTag}
+          onChange={(e) => setSelectedTag(e.target.value)}
+          className="p-2 border rounded-md bg-white shadow-sm"
+        >
+          <option value="ALL">全て</option>
+          {tags.map((tag) => (
+            <option key={tag} value={tag}>
+              {tag}
+            </option>
+          ))}
+        </select>
       </div>
 
-      <div className="space-y-12">
+      <div className="space-y-12 mt-10">
         {Object.entries(grouped).map(([ym, bms]) => (
           <section key={ym}>
-            <div data-ym={ym} className="flex justify-center">
+            <div className="flex justify-center mb-4">
               <YearMonthCard label={ym} />
             </div>
-            <div className="mt-8 grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6">
-              {bms.map((bk) => {
-                const info = tagMap[bk.id] ?? { tagListId: '', tags: [] };
-                return (
-                  <BookmarkCard
-                    key={bk.id}
-                    bookmark={bk}
-                    editable={editable}
-                    servicePath={servicePath}
-                    initialTags={info.tags}
-                    initialTagListId={info.tagListId}
-                  />
-                );
-              })}
+            <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6">
+              {bms.map((bk) => (
+                <BookmarkCard
+                  key={bk.id}
+                  bookmark={bk}
+                  editable={editable}
+                  servicePath={servicePath}
+                  initialTags={bookmarkTagDataMap[bk.id].tags}
+                  initialTagListId={bookmarkTagDataMap[bk.id].tagListId}
+                />
+              ))}
             </div>
           </section>
         ))}
